user dashboard

+++ app/Http/Controllers/DashboardController.php
@@ -29,6 +29,9 @@ class DashboardController extends Controller
 
     public function dashboard(){
         $user = User::where('id', Auth::id())->first();
+        if ($user->type !='member'){
+            return redirect('dashboard/timeline');
+        }
         $blogids = Blog::where('user_id', Auth::id())->select('id')->get();
         // $comments = Comment::whereIn('blog_id', $blogids)->where('parent_id',NULL)->orderBy('id', 'desc')->take(3)->get();
         $comments = DB::table('comments')->whereIn('blog_id', $blogids->pluck('id'))->where('parent_id', NULL)
@@ -52,24 +55,37 @@ class DashboardController extends Controller
 
     public function blogs(){
         $user = User::where('id', Auth::id())->first();
+        if ($user->type !='member'){
+            return redirect('dashboard/timeline');
+        }
         $blogs  = Blog::where('user_id', Auth::id())->with('user', 'comments')->orderBy('id', 'desc')->paginate(21);
         return view('dashboard.blog', compact('user','blogs'));
     }
     
     public function newblog(){
         $user = User::where('id', Auth::id())->first();
+        if ($user->type !='member'){
+            return redirect('dashboard/timeline');
+        }
         $blogs  = Blog::where('user_id', Auth::id())->with('user', 'comments')->orderBy('id', 'desc')->paginate(21);
         return view('dashboard.newblog', compact('user','blogs'));
     }
 
     public function editblog(Request $request){
         $user = User::where('id', Auth::id())->first();
+        if ($user->type !='member'){
+            return redirect('dashboard/timeline');
+        }
         $blog  = Blog::where('user_id', Auth::id())->where('slug', $request->slug)->with('user', 'comments')->orderBy('id', 'desc')->first();
         return view('dashboard.editblog', compact('user','blog'));
     }
 
     public function comments(){
         $user = User::where('id', Auth::id())->first();
+        if ($user->type !='member'){
+            $comments = Comment::where('user_id', Auth::id())->with('blog')->paginate(18);
+            return view('dashboard.comments', compact('user','comments'));
+        }
         $blogids = Blog::where('user_id', Auth::id())->select('id')->get();
         $comments = DB::table('comments')->whereIn('blog_id', $blogids->pluck('id'))->where('parent_id', NULL)
             ->join('users', 'comments.user_id', '=', 'users.id')
@@ -77,13 +93,15 @@ class DashboardController extends Controller
             ->leftjoin('flags', 'comments.id', '=', 'flags.subject_id')
             ->select('comments.*', 'users.name', 'users.avatar', 'blogs.slug', 'blogs.title',DB::raw('IF(STRCMP(flags.type,"comment") = 0, TRUE, FALSE) as flagged'))
             ->orderBy('id', 'desc')->paginate(18);
-        // return $comments = Comment::whereIn('blog_id', $blogids->pluck('id'))->where('parent_id', NULL)->with('user')->orderBy('id', 'desc')->paginate(18);
         return view('dashboard.comments', compact('user','comments'));
     }
 
 
     public function reviews(){
         $user = User::where('id', Auth::id())->first();
+        if ($user->type !='member'){
+            return redirect('dashboard/timeline');
+        }
         $blogids = Blog::where('user_id', Auth::id())->select('id')->get();
         $reviews = DB::table('reviews')->where('reviews.member_id', Auth::id())
             ->join('users', 'reviews.user_id', '=', 'users.id')
@@ -97,14 +115,21 @@ class DashboardController extends Controller
 
     public function questions(){
         $user = User::where('id', Auth::id())->first();
+        if ($user->type !='member'){
+            $questions = Question::with('user')->where('user_id', $user->id)->paginate(18);
+            return view('dashboard.questions', compact('user','questions'));
+        }
         $questions = Question::with('user')->where('member_id', $user->id)->where('answers_count',0)->paginate(18);
-        // return $answers;
         return view('dashboard.questions', compact('user','questions'));
     }
 
 
     public function answers(){
         $user = User::where('id', Auth::id())->first();
+        if ($user->type !='member'){
+            // TODO
+            return redirect('dashboard/timeline');
+        }
         $answers = Answer::with('user', 'question')->where('user_id', $user->id)->paginate(18);
         // return $answers;
         return view('dashboard.answers', compact('user','answers'));
@@ -112,10 +137,22 @@ class DashboardController extends Controller
 
     public function followers(){
         $user = User::where('id', Auth::id())->first();
+        if ($user->type !='member'){
+            return redirect('dashboard/timeline');
+        }
         $followers = $user->followers;
         return view('dashboard.followers', compact('user','followers'));
     }
 
+    public function followings(){
+        $user = User::where('id', Auth::id())->first();
+        if ($user->type !='user'){
+            return redirect('dashboard');
+        }
+        $followings = $user->followings;
+        return view('dashboard.followings', compact('user','followings'));
+    }
+    
     public function profile(){
         $user = User::where('id', Auth::id())->first();
         $countries = Country::where('is_active', 1)->get();
@@ -174,6 +211,13 @@ class DashboardController extends Controller
         return view('dashboard.messages', compact('user'));
     }
 
+    public function timeline(){
+        $user = User::where('id', Auth::id())->first();
+        $followers =  $user->followings->take(12);
+        $activites = Activity::whereIn('causer_id',$followers->pluck('id'))->where('log_name','timeline')->orderBy('id','desc')->paginate(20);
+        return view('dashboard.timeline', compact('user','activites'));
+    }
+
     public function editorupload(Request $request)
     {
         if($request->hasFile('upload')) {
@@ -196,6 +240,10 @@ class DashboardController extends Controller
 
     public function storeMedia(Request $request)
     {
+        $user = User::where('id', Auth::id())->first();
+        if ($user->type !='member'){
+            return redirect('dashboard/timeline');
+        }
         $path = storage_path('tmp/uploads');
     
         if (!file_exists($path)) {
@@ -203,9 +251,7 @@ class DashboardController extends Controller
         }
         $file = $request->file('file');
         $uploadedimage = $request->file->store('tmp','public');
-
-        // $name = uniqid() . '_' . trim($file->getClientOriginalName());
-        // $file->move($path, $name);
+    
         return response()->json([
             'name'          => $uploadedimage,
             'original_name' => $file->getClientOriginalName(),
@@ -215,18 +261,31 @@ class DashboardController extends Controller
     public function storeimage(Request $request)
     {
         $user = User::where('id', Auth::id())->first();
-        foreach ($request->input('images', []) as $file) {
+        if ($user->type !='member'){
+            return redirect('dashboard/timeline');
+        }        foreach ($request->input('images', []) as $file) {
             $user->addMedia(public_path('/storage/'.$file))->toMediaCollection('photos');
         }
-        return response()->json([
-            'name'          => "stored",
-        ]);
+            activity('timeline')
+            ->causedBy(Auth::id())
+            ->performedOn($user)
+            ->withProperties([
+                'slug' => '/member/'.Auth::user()->slug,
+                'useravatar' => Auth::user()->avatar,
+                'username' => Auth::user()->name,
+                'userslug' => Auth::user()->slug
+                ])
+            ->log(':causer.name has updated picture(s) '); 
+        $arr = array('msg' => 'Successfully stored', 'status' => true);
+        return Response()->json($arr);
     }
 
     public function deletemedia(Request $request, User $user)
     {
         $user = User::where('id', Auth::id())->first();
-        $photos = $user->getMedia('photos');
+        if ($user->type !='member'){
+            return redirect('dashboard/timeline');
+        }        $photos = $user->getMedia('photos');
         foreach ($photos as $photo){
             if ($photo->id == $request->pic && $photo->model_id == Auth::id()){
                 $photo->delete();
@@ -243,21 +302,34 @@ class DashboardController extends Controller
     public function storevideo(Request $request)
     {
         $user = User::where('id', Auth::id())->first();
-        foreach ($request->input('videos', []) as $file) {
+        if ($user->type !='member'){
+            return redirect('dashboard/timeline');
+        }        foreach ($request->input('videos', []) as $file) {
             $user->addMedia(public_path('/storage/'.$file))->toMediaCollection('videos');
         }
-        return response()->json([
-            'name'          => "stored",
-        ]);
+            activity('timeline')
+            ->causedBy(Auth::id())
+            ->performedOn($user)
+            ->withProperties([
+                'slug' => '/member/'.Auth::user()->slug,
+                'useravatar' => Auth::user()->avatar,
+                'username' => Auth::user()->name,
+                'userslug' => Auth::user()->slug
+                ])
+            ->log(':causer.name has updated video(s)'); 
+        $arr = array('msg' => 'Successfully stored', 'status' => true);
+        return Response()->json($arr);
     }
 
     public function deletevideo(Request $request, User $user)
     {
         $user = User::where('id', Auth::id())->first();
-        $photos = $user->getMedia('photos');
-        foreach ($photos as $photo){
-            if ($photo->id == $request->pic && $photo->model_id == Auth::id()){
-                $photo->delete();
+        if ($user->type !='member'){
+            return redirect('dashboard/timeline');
+        }        $videos = $user->getMedia('videos');
+        foreach ($videos as $video){
+            if ($video->id == $request->vid && $video->model_id == Auth::id()){
+                $video->delete();
                 return response()->json([
                     'msg'          => "success",
                 ]);
@@ -273,19 +345,20 @@ class DashboardController extends Controller
         $file = $request->file('file');
         $user = User::where('id', Auth::id())->first();
         $uploadedimage = $request->file->store('users-avatar','public');
-        // https://randomuser.me/api/portraits/men/1.jpg
-        // return response()->json([
-        //     'name'          => $uploadedimage,
-        //     'original_name' => $file->getClientOriginalName(),
-        // ]);
         $userupdate = User::where('id', Auth::id())
                         ->update(['avatar' => $uploadedimage]);
             $arr = array('msg' => 'Something goes to wrong. Please try again later', 'status' => false);
-            if($userupdate){ 
-                activity()
+            if($userupdate){
+                activity('timeline')
                 ->causedBy(Auth::id())
                 ->performedOn($user)
-                ->log(':causer.name Updated Profile picture');
+                ->withProperties([
+                    'slug' => '/member/'.Auth::user()->slug,
+                    'useravatar' => Auth::user()->avatar,
+                    'username' => Auth::user()->name,
+                    'userslug' => Auth::user()->slug
+                    ])
+                ->log(':causer.name has updated profile picture '); 
             $arr = array('msg' => 'Successfully stored', 'status' => true);
             }
             return Response()->json($arr);
